Added cast to get MathTest working on Windows
git-svn-id: http://skia.googlecode.com/svn/trunk@4072 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ tests/MathTest.cpp
@@ -19,7 +19,7 @@ static float sk_fsel(float pred, float result_ge, float result_lt) {
 static float fast_floor(float x) {
 //    float big = sk_fsel(x, 0x1.0p+23, -0x1.0p+23);
     float big = sk_fsel(x, (float)(1 << 23), -(float)(1 << 23));
-    return (x + big) - big;
+    return (float)(x + big) - big;
 }
 
 static float std_floor(float x) {
